出品者のみ削除可能

+++ app/controllers/items_controller.rb
@@ -25,7 +25,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if user_signed_in? && current_user.id == item.user_id
     redirect_to root_path
   end
